le tengo que dar una vuelta a los permisos xd

+++ routes/api.php
@@ -11,6 +11,8 @@ use App\Http\Controllers\api\PedidoController;
 use App\Http\Controllers\api\LineaController;
 use App\Http\Controllers\api\FacturaController;
 use App\Http\Middleware\AdminCheck;
+use App\Http\Middleware\MeseroCheck;
+use App\Http\Middleware\UserIsOwnerCheck;
 use App\Http\Middleware\UserNotBlockedCheck;
 use Illuminate\Support\Facades\Route;
 
@@ -45,12 +47,12 @@ Route::group(['middleware' => ['auth:sanctum', UserNotBlockedCheck::class]], fun
      *  5. Editar un usuario
      *  6. Eliminar un usuario
      */
-    Route::get('/usuarios', [UserController::class, 'index']);
-    Route::get('/usuarios/pages', [UserController::class, 'getAmountOfpages']);
-    Route::get('/usuarios/{id}', [UserController::class, 'getUser']);
-    Route::get('/usuarios/{id}/pedidos', [UserController::class, 'getUsersPedidos']);
-    Route::put('/usuarios/{id}', [UserController::class, 'updateUser']);
-    Route::delete('/usuarios/{id}', [UserController::class, 'deleteUser']);
+    Route::get('/usuarios', [UserController::class, 'index'])->middleware([AdminCheck::class]);
+    Route::get('/usuarios/pages', [UserController::class, 'getAmountOfpages'])->middleware([AdminCheck::class]);
+    Route::get('/usuarios/{id}', [UserController::class, 'getUser']); // TODO: mirar los permisos para esta ruta
+    Route::get('/usuarios/{id}/pedidos', [UserController::class, 'getUsersPedidos']); // TODO: mirar los permisos para esta ruta
+    Route::put('/usuarios/{id}', [UserController::class, 'updateUser'])->middleware([UserIsOwnerCheck::class]);
+    Route::delete('/usuarios/{id}', [UserController::class, 'deleteUser'])->middleware([AdminCheck::class]);
 
     /**
      *  Categorias endpoints
@@ -62,13 +64,13 @@ Route::group(['middleware' => ['auth:sanctum', UserNotBlockedCheck::class]], fun
      *  6. Todos los productos de una categoría (ID)
      *  7. Editar una categoría (ID)
      */
-    Route::get('/categorias', [CategoriaController::class, 'index']);
-    Route::get('/categorias/pages', [CategoriaController::class, 'getAmountOfPages']);
-    Route::get('/categorias/{id}', [CategoriaController::class, 'getCategoria']);
-    Route::post('/categorias/new', [CategoriaController::class, 'newCategoria']);
-    Route::delete('/categorias/{id}', [CategoriaController::class, 'deleteCategoria']);
-    Route::get('/categorias/{id}/productos', [ProductoController::class, 'getProductosByCategoria']);
-    Route::put('/categorias/{id}', [CategoriaController::class, 'updateCategoria']);
+    Route::get('/categorias', [CategoriaController::class, 'index'])->middleware([AdminCheck::class]);
+    Route::get('/categorias/pages', [CategoriaController::class, 'getAmountOfPages'])->middleware([AdminCheck::class]);
+    Route::get('/categorias/{id}', [CategoriaController::class, 'getCategoria']); // TODO: mirar los permisos para esta ruta
+    Route::post('/categorias/new', [CategoriaController::class, 'newCategoria'])->middleware([AdminCheck::class]);
+    Route::delete('/categorias/{id}', [CategoriaController::class, 'deleteCategoria'])->middleware([AdminCheck::class]);
+    Route::get('/categorias/{id}/productos', [ProductoController::class, 'getProductosByCategoria']); // TODO: mirar los permisos para esta ruta
+    Route::put('/categorias/{id}', [CategoriaController::class, 'updateCategoria']); // TODO: mirar los permisos para esta ruta
 
     /**
      *  Productos endpoints
@@ -82,15 +84,15 @@ Route::group(['middleware' => ['auth:sanctum', UserNotBlockedCheck::class]], fun
      *  8. Aumentar el stock de un producto
      *  9. Reducir el stock de un producto
      */
-    Route::get('/productos', [ProductoController::class, 'index']);
-    Route::get('/productos/pages', [ProductoController::class, 'getAmountOfPages']);
-    Route::get('/productos/{id}', [ProductoController::class, 'getProducto']);
-    Route::post('/productos/new', [ProductoController::class, 'newProducto']);
-    Route::delete('/productos/{id}', [ProductoController::class, 'deleteProducto']);
-    Route::put('/productos/{id}', [ProductoController::class, 'updateProducto']);
-    Route::get('/productos/{id}/stock', [ProductoController::class, 'getProductStock']);
-    Route::post('/productos/{id}/stock/add', [ProductoController::class, 'addStock']);
-    Route::post('/productos/{id}/stock/reduce', [ProductoController::class, 'reduceStock']);
+    Route::get('/productos', [ProductoController::class, 'index']); // TODO: mirar los permisos para esta ruta
+    Route::get('/productos/pages', [ProductoController::class, 'getAmountOfPages']); // TODO: mirar los permisos para esta ruta
+    Route::get('/productos/{id}', [ProductoController::class, 'getProducto']); // TODO: mirar los permisos para esta ruta
+    Route::post('/productos/new', [ProductoController::class, 'newProducto']); // TODO: mirar los permisos para esta ruta
+    Route::delete('/productos/{id}', [ProductoController::class, 'deleteProducto'])->middleware([AdminCheck::class]);
+    Route::put('/productos/{id}', [ProductoController::class, 'updateProducto']); // TODO: mirar los permisos para esta ruta
+    Route::get('/productos/{id}/stock', [ProductoController::class, 'getProductStock']); // TODO: mirar los permisos para esta ruta
+    Route::post('/productos/{id}/stock/add', [ProductoController::class, 'addStock']); // TODO: mirar los permisos para esta ruta
+    Route::post('/productos/{id}/stock/reduce', [ProductoController::class, 'reduceStock']); // TODO: mirar los permisos para esta ruta
 
     /**
      *  Stock endpoints
@@ -99,10 +101,10 @@ Route::group(['middleware' => ['auth:sanctum', UserNotBlockedCheck::class]], fun
      *  2. Dar de alta un producto en stock (Dudo que se vaya a utilizar)
      *  3. Editar un stock
      */
-    Route::get('/stock', [StockController::class, 'index']);
-    Route::get('/stock/pages', [StockController::class, 'getAmountOfPages']);
-    Route::post('/stock/new', [StockController::class, 'createStock']);
-    Route::put('/stock/{id}', [StockController::class, 'updateStock']);
+    Route::get('/stock', [StockController::class, 'index']); // TODO: mirar los permisos para esta ruta
+    Route::get('/stock/pages', [StockController::class, 'getAmountOfPages']); // TODO: mirar los permisos para esta ruta
+    Route::post('/stock/new', [StockController::class, 'createStock']); // TODO: mirar los permisos para esta ruta
+    Route::put('/stock/{id}', [StockController::class, 'updateStock']); // TODO: mirar los permisos para esta ruta
 
     /**
      *  Mesas endpoints
@@ -114,13 +116,13 @@ Route::group(['middleware' => ['auth:sanctum', UserNotBlockedCheck::class]], fun
      *  6. Listar todos los pedidos de una mesa
      *  7. Listar el pedido actual de una mesa
      */
-    Route::get('/mesas', [MesaController::class, 'index']);
-    Route::get('/mesas/{id}', [MesaController::class, 'getMesa']);
-    Route::post('/mesas/new', [MesaController::class, 'newMesa']);
-    Route::delete('/mesas/{id}', [MesaController::class, 'deleteMesa']);
-    Route::put('/mesas/{id}', [MesaController::class, 'updateMesa']);
-    Route::get('mesas/{id}/pedidos', [MesaController::class, 'getPedidosByMesa']);
-    Route::get('/mesas/{id}/pedido', [MesaController::class, 'getPedidoActual']);
+    Route::get('/mesas', [MesaController::class, 'index']); // TODO: mirar los permisos para esta ruta
+    Route::get('/mesas/{id}', [MesaController::class, 'getMesa']); // TODO: mirar los permisos para esta ruta
+    Route::post('/mesas/new', [MesaController::class, 'newMesa'])->middleware([AdminCheck::class]);
+    Route::delete('/mesas/{id}', [MesaController::class, 'deleteMesa'])->middleware([AdminCheck::class]);
+    Route::put('/mesas/{id}', [MesaController::class, 'updateMesa']); // TODO: mirar los permisos para esta ruta
+    Route::get('mesas/{id}/pedidos', [MesaController::class, 'getPedidosByMesa']); // TODO: mirar los permisos para esta ruta
+    Route::get('/mesas/{id}/pedido', [MesaController::class, 'getPedidoActual']); // TODO: mirar los permisos para esta ruta
 
     /**
      *  Pedidos endpoints
@@ -134,15 +136,15 @@ Route::group(['middleware' => ['auth:sanctum', UserNotBlockedCheck::class]], fun
      *  8. Buscar la factura de un pedido
      *  9. Cambiar el estado de un pedido a 'servido'
      */
-    Route::get('/pedidos', [PedidoController::class, 'index']);
-    Route::get('/pedidos/pages', [PedidoController::class, 'getAmountOfPages']);
-    Route::get('/pedidos/{id}', [PedidoController::class, 'getPedido']);
-    Route::post('/pedidos/new', [PedidoController::class, 'newPedido'])->middleware([\App\Http\Middleware\MeseroCheck::class]);
-    Route::put('/pedidos/{id}', [PedidoController::class, 'updatePedido']);
-    Route::delete('/pedidos/{id}', [PedidoController::class, 'deletePedido']);
-    Route::get('/pedidos/{id}/lineas', [LineaController::class, 'getLineasByPedido']);
-    Route::get('/pedidos/{id}/factura', [FacturaController::class, 'getFacturaByPedido']);
-    Route::post('/pedidos/{id}/servir', [PedidoController::class, 'servirPedido']);
+    Route::get('/pedidos', [PedidoController::class, 'index']); // TODO: mirar los permisos para esta ruta
+    Route::get('/pedidos/pages', [PedidoController::class, 'getAmountOfPages']); // TODO: mirar los permisos para esta ruta
+    Route::get('/pedidos/{id}', [PedidoController::class, 'getPedido']); // TODO: mirar los permisos para esta ruta
+    Route::post('/pedidos/new', [PedidoController::class, 'newPedido'])->middleware([MeseroCheck::class]);
+    Route::put('/pedidos/{id}', [PedidoController::class, 'updatePedido']); // TODO: mirar los permisos para esta ruta
+    Route::delete('/pedidos/{id}', [PedidoController::class, 'deletePedido']); // TODO: mirar los permisos para esta ruta
+    Route::get('/pedidos/{id}/lineas', [LineaController::class, 'getLineasByPedido']); // TODO: mirar los permisos para esta ruta
+    Route::get('/pedidos/{id}/factura', [FacturaController::class, 'getFacturaByPedido']); // TODO: mirar los permisos para esta ruta
+    Route::post('/pedidos/{id}/servir', [PedidoController::class, 'servirPedido']); // TODO: mirar los permisos para esta ruta
 
     /**
      *  Líneas endpoints
@@ -153,12 +155,12 @@ Route::group(['middleware' => ['auth:sanctum', UserNotBlockedCheck::class]], fun
      *  5. Modificar una línea
      *  6. Eliminar una línea
      */
-    Route::get('/lineas', [LineaController::class, 'index']);
-    Route::get('/lineas/pages', [LineaController::class, 'getAmountOfPages']);
-    Route::get('/lineas/{id}', [LineaController::class, 'getLinea']);
-    Route::post('/lineas/new', [LineaController::class, 'newLinea']);
-    Route::put('/lineas/{id}', [LineaController::class, 'updateLinea']);
-    Route::delete('/lineas/{id}', [LineaController::class, 'deleteLinea']);
+    Route::get('/lineas', [LineaController::class, 'index']); // TODO: mirar los permisos para esta ruta
+    Route::get('/lineas/pages', [LineaController::class, 'getAmountOfPages']); // TODO: mirar los permisos para esta ruta
+    Route::get('/lineas/{id}', [LineaController::class, 'getLinea']); // TODO: mirar los permisos para esta ruta
+    Route::post('/lineas/new', [LineaController::class, 'newLinea']); // TODO: mirar los permisos para esta ruta
+    Route::put('/lineas/{id}', [LineaController::class, 'updateLinea']); // TODO: mirar los permisos para esta ruta
+    Route::delete('/lineas/{id}', [LineaController::class, 'deleteLinea']); // TODO: mirar los permisos para esta ruta
 
     /**
      *  Facturas endpoints
@@ -169,11 +171,11 @@ Route::group(['middleware' => ['auth:sanctum', UserNotBlockedCheck::class]], fun
      *  4. Modificar una factura
      *  5. Eliminar una factura
      */
-    Route::get('/facturas', [FacturaController::class, 'index']);
-    Route::get('/facturas/pages', [FacturaController::class, 'getAmountOfPages']);
-    Route::get('/facturas/{id}', [FacturaController::class, 'getFactura']);
-    Route::post('/facturas/new', [FacturaController::class, 'newFactura']);
-    Route::put('/facturas/{id}', [FacturaController::class, 'updateFactura']);
-    Route::delete('/facturas/{id}', [FacturaController::class, 'deleteFactura']);
+    Route::get('/facturas', [FacturaController::class, 'index']); // TODO: mirar los permisos para esta ruta
+    Route::get('/facturas/pages', [FacturaController::class, 'getAmountOfPages']); // TODO: mirar los permisos para esta ruta
+    Route::get('/facturas/{id}', [FacturaController::class, 'getFactura']); // TODO: mirar los permisos para esta ruta
+    Route::post('/facturas/new', [FacturaController::class, 'newFactura']); // TODO: mirar los permisos para esta ruta
+    Route::put('/facturas/{id}', [FacturaController::class, 'updateFactura']); // TODO: mirar los permisos para esta ruta
+    Route::delete('/facturas/{id}', [FacturaController::class, 'deleteFactura'])->middleware([AdminCheck::class]);
 });
